pyproject.toml
[project]
name = "sabs-resume"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "diskcache>=5.6.3",
    "fastapi>=0.110.0",
    "google-genai",
    "pandas>=2.0.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "python-multipart>=0.0.9",
    "structlog>=24.1.0",
    "tenacity>=8.2.3",
    "uvicorn[standard]>=0.27.0",
]
